HOTFIX Agrega al archivo travis la variable de entorno para testing de Cloudinary. Agrega log.

+++ src/main/java/sic/service/impl/PhotoVideoUploaderImpl.java
@@ -25,6 +25,9 @@ public class PhotoVideoUploaderImpl implements IPhotoVideoUploader {
   @Value("${CLOUDINARY_URL}")
   private String cloudinaryUrl;
 
+  @Value("${SIC_CLOUDINARY_ENV}")
+  private String cloudinaryEnv;
+
   private final Logger logger = LoggerFactory.getLogger(this.getClass());
   private final MessageSource messageSource;
 
@@ -35,34 +38,51 @@ public class PhotoVideoUploaderImpl implements IPhotoVideoUploader {
 
   @Override
   public String subirImagen(String nombreImagen, byte[] imagen) {
-    String urlImagen;
-    try {
-      Path path = Files.createTempFile("imagen-file", ".jpg");
-      File file = path.toFile();
-      Files.write(path, imagen);
-      Cloudinary cloudinary = new Cloudinary(cloudinaryUrl);
-      Map uploadResult = cloudinary.uploader().upload(file, ObjectUtils.asMap(
-              "public_id", nombreImagen,
-              "transformation", new Transformation().crop("fit").width(800).height(600)));
-      urlImagen = uploadResult.get("secure_url").toString();
-      logger.warn("La imagen {} se guardó correctamente.", nombreImagen);
-      Files.delete(path);
-    } catch (IOException ex) {
-      throw new ServiceException(messageSource.getMessage(
-        "mensaje_error_al_subir_imagen", null, Locale.getDefault()), ex);
+    String urlImagen = "";
+    if (cloudinaryEnv.equals("production")) {
+      try {
+        Path path = Files.createTempFile("imagen-file", ".jpg");
+        File file = path.toFile();
+        Files.write(path, imagen);
+        Cloudinary cloudinary = new Cloudinary(cloudinaryUrl);
+        Map uploadResult =
+            cloudinary
+                .uploader()
+                .upload(
+                    file,
+                    ObjectUtils.asMap(
+                        "public_id",
+                        nombreImagen,
+                        "transformation",
+                        new Transformation().crop("fit").width(800).height(600)));
+        urlImagen = uploadResult.get("secure_url").toString();
+        logger.warn("La imagen {} se guardó correctamente.", nombreImagen);
+        Files.delete(path);
+      } catch (IOException ex) {
+        throw new ServiceException(
+            messageSource.getMessage("mensaje_error_al_subir_imagen", null, Locale.getDefault()),
+            ex);
+      }
+    } else {
+      logger.error("Cloudinary environment = {}, la imagen no se procesó.", cloudinaryEnv);
     }
     return urlImagen;
   }
 
   @Override
   public void borrarImagen(String publicId) {
-    try {
-      Cloudinary cloudinary = new Cloudinary(cloudinaryUrl);
-      cloudinary.uploader().destroy(publicId, ObjectUtils.emptyMap());
-      logger.warn("La imagen {} se eliminó correctamente.", publicId);
-    } catch (IOException ex) {
-      throw new ServiceException(messageSource.getMessage(
-        "mensaje_error_al_borrar_imagen", null, Locale.getDefault()), ex);
+    if (cloudinaryEnv.equals("production")) {
+      try {
+        Cloudinary cloudinary = new Cloudinary(cloudinaryUrl);
+        cloudinary.uploader().destroy(publicId, ObjectUtils.emptyMap());
+        logger.warn("La imagen {} se eliminó correctamente.", publicId);
+      } catch (IOException ex) {
+        throw new ServiceException(
+            messageSource.getMessage("mensaje_error_al_borrar_imagen", null, Locale.getDefault()),
+            ex);
+      }
+    } else {
+      logger.error("Cloudinary environment = {}, la imagen no se eliminó.", cloudinaryEnv);
     }
   }
 }
